backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
llama-index
langchain
openai
pypdf
python-dotenv 